+++ src/NodeJS/Javascript/Servers/OutOfProcess/Http/Shared.ts
@@ -1,4 +1,5 @@
 ﻿const path = require("path");
+import * as stream from 'stream';
 import * as http from 'http';
 import * as http2 from 'http2';
 import InvocationRequest from "../../../InvocationData/InvocationRequest";